Added ability to retrieve ID of last bucket row inserted (now passed back to user) -- (updated to ExecuteScalar & query to slqite_sequence --  & also passing back JSON object with mainTokenId & bucketId so user can retrieve this specific bucket data later.

+++ LibreStore/Interfaces/IPersistable.cs
@@ -1,4 +1,4 @@
 public interface IPersistable
 {
-	int Save();
+	Int64 Save();
 }
